requirements.txt
tensorflow==2.15.0
streamlit
opencv-python
pillow
numpy
